relocate webhook-collector and generalize, closes #6
