+++ .github/workflows/ci-webhook-collector.yml
@@ -1,0 +1,43 @@
+name: CI Webhook Collector
+
+on:
+  pull_request:
+    paths:
+      - 'webhook-collector/**/*'
+
+jobs:
+  build-and-test:
+    runs-on: ubuntu-latest
+    env:
+      TARGET_TYPE: opensearch
+      OPENSEARCH_HOST: opensearch-node1
+      OPENSEARCH_PORT: 9200
+      OPENSEARCH_PROTOCOL: 'https'
+      OPENSEARCH_USERNAME: 'admin'
+      OPENSEARCH_PASSWORD: 'ChangeMe123!'
+      OPENSEARCH_VERSION: '2.14.0'
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20'
+      - name: Install Dependencies
+        run: |
+          cd webhook-collector
+          npm ci
+      - name: Run Tests
+        run: |
+          cd webhook-collector
+          npm test
+      - name: Check if Dist is up to date
+        run: |
+          cd webhook-collector
+          npm run build
+          if [ "$(git diff --ignore-space-at-eol dist/ | wc -l)" -gt "0" ]; then
+            echo "Detected uncommitted changes after build.  See status below:"
+            git diff
+            exit 1
+          fi
+        id: diff
