+++ grafana-opensearch/.env.example
@@ -1,3 +1,4 @@
+TARGET_TYPE=opensearch
 OPENSEARCH_HOST=opensearch-node1
 OPENSEARCH_PORT=9200
 OPENSEARCH_PROTOCOL=https
