+++ webhook-collector/.env.example
@@ -1,0 +1,8 @@
+export TARGET_TYPE=opensearch
+export OPENSEARCH_HOST=opensearch-node1
+export OPENSEARCH_PORT=9200
+export OPENSEARCH_PROTOCOL=https
+export OPENSEARCH_USERNAME=admin
+export OPENSEARCH_PASSWORD=ChangeMe123!
+export OPENSEARCH_DASHBOARD_PORT=5601
+export OPENSEARCH_VERSION=2.14.0
